+++ src/pages/GameSection.jsx
@@ -1,7 +1,6 @@
 import { Link } from "react-router-dom";
 import { motion } from "framer-motion";
 import Navbar from "../components/navbar"
-import Footer from "../components/footer"
 export default function GameSection() {
   const games = [
     {
@@ -66,7 +65,7 @@ export default function GameSection() {
         </div>
       </div>
     </div>
-    <Footer/>
+   
     </>
   );
 }
